move connector http transport structs into apollo-federation
