+++ apollo-federation/Cargo.toml
@@ -48,6 +48,9 @@ either = "1.13.0"
 tracing = "0.1.40"
 ron = { version = "0.10.0", optional = true }
 shape = "0.5.2"
+form_urlencoded = "1.2.1"
+parking_lot = "0.12.4"
+mime = "0.3.17"
 
 [dev-dependencies]
 diff = "0.1.13"
