+++ apollo-federation/src/connectors/runtime.rs
@@ -1,3 +1,5 @@
 pub mod debug;
+pub mod form_encoding;
+pub mod http_json_transport;
 pub mod inputs;
 pub mod mapping;
