+++ apollo-router/src/plugins/connectors/make_requests.rs
@@ -8,11 +8,11 @@ use apollo_federation::connectors::Connector;
 use apollo_federation::connectors::EntityResolver;
 use apollo_federation::connectors::JSONSelection;
 use apollo_federation::connectors::runtime::debug::ConnectorContext;
+use apollo_federation::connectors::runtime::http_json_transport::HttpJsonTransportError;
+use apollo_federation::connectors::runtime::http_json_transport::make_request;
 use apollo_federation::connectors::runtime::inputs::RequestInputs;
 use parking_lot::Mutex;
 
-use super::http_json_transport::HttpJsonTransportError;
-use super::http_json_transport::make_request;
 use crate::Context;
 use crate::json_ext::Path;
 use crate::json_ext::PathElement;
@@ -671,13 +671,13 @@ mod tests {
     use apollo_federation::connectors::Connector;
     use apollo_federation::connectors::HttpJsonTransport;
     use apollo_federation::connectors::JSONSelection;
+    use apollo_federation::connectors::runtime::http_json_transport::TransportRequest;
     use http::Uri;
     use insta::assert_debug_snapshot;
 
     use crate::Context;
     use crate::graphql;
     use crate::query_planner::fetch::Variables;
-    use crate::services::connector::request_service::TransportRequest;
 
     #[test]
     fn test_root_fields_simple() {
