+++ apollo-router/src/plugins/connectors/mod.rs
@@ -1,7 +1,5 @@
 pub(crate) mod configuration;
-mod form_encoding;
 pub(crate) mod handle_responses;
-pub(crate) mod http_json_transport;
 pub(crate) mod incompatible;
 pub(crate) mod make_requests;
 pub(crate) mod plugin;
