+++ apollo-router/src/plugins/telemetry/config_new/connector/events.rs
@@ -136,6 +136,10 @@ mod tests {
     use apollo_federation::connectors::JSONSelection;
     use apollo_federation::connectors::SourceName;
     use apollo_federation::connectors::StringTemplate;
+    use apollo_federation::connectors::runtime::http_json_transport::HttpRequest;
+    use apollo_federation::connectors::runtime::http_json_transport::HttpResponse;
+    use apollo_federation::connectors::runtime::http_json_transport::TransportRequest;
+    use apollo_federation::connectors::runtime::http_json_transport::TransportResponse;
     use http::HeaderValue;
     use tracing::instrument::WithSubscriber;
 
@@ -147,9 +151,6 @@ mod tests {
     use crate::plugins::test::PluginTestHarness;
     use crate::services::connector::request_service::Request;
     use crate::services::connector::request_service::Response;
-    use crate::services::connector::request_service::TransportRequest;
-    use crate::services::connector::request_service::TransportResponse;
-    use crate::services::connector::request_service::transport;
     use crate::services::router::body;
 
     #[tokio::test(flavor = "multi_thread")]
@@ -166,7 +167,7 @@ mod tests {
             http_request
                 .headers_mut()
                 .insert("x-log-request", HeaderValue::from_static("log"));
-            let transport_request = TransportRequest::Http(transport::http::HttpRequest {
+            let transport_request = TransportRequest::Http(HttpRequest {
                 inner: http_request,
                 debug: Default::default(),
             });
@@ -215,7 +216,7 @@ mod tests {
                 .call_connector_request_service(connector_request, |request| Response {
                     context: request.context.clone(),
                     connector: request.connector.clone(),
-                    transport_result: Ok(TransportResponse::Http(transport::http::HttpResponse {
+                    transport_result: Ok(TransportResponse::Http(HttpResponse {
                         inner: http::Response::builder()
                             .status(200)
                             .header("x-log-request", HeaderValue::from_static("log"))
@@ -253,7 +254,7 @@ mod tests {
             http_request
                 .headers_mut()
                 .insert("x-log-response", HeaderValue::from_static("log"));
-            let transport_request = TransportRequest::Http(transport::http::HttpRequest {
+            let transport_request = TransportRequest::Http(HttpRequest {
                 inner: http_request,
                 debug: Default::default(),
             });
@@ -302,7 +303,7 @@ mod tests {
                 .call_connector_request_service(connector_request, |request| Response {
                     context: request.context.clone(),
                     connector: request.connector.clone(),
-                    transport_result: Ok(TransportResponse::Http(transport::http::HttpResponse {
+                    transport_result: Ok(TransportResponse::Http(HttpResponse {
                         inner: http::Response::builder()
                             .status(200)
                             .header("x-log-response", HeaderValue::from_static("log"))
