+++ apollo-router/src/plugins/telemetry/config_new/instruments.rs
@@ -2452,6 +2452,10 @@ mod tests {
     use apollo_federation::connectors::JSONSelection;
     use apollo_federation::connectors::SourceName;
     use apollo_federation::connectors::StringTemplate;
+    use apollo_federation::connectors::runtime::http_json_transport::HttpRequest;
+    use apollo_federation::connectors::runtime::http_json_transport::HttpResponse;
+    use apollo_federation::connectors::runtime::http_json_transport::TransportRequest;
+    use apollo_federation::connectors::runtime::http_json_transport::TransportResponse;
     use apollo_federation::connectors::runtime::mapping::Problem;
     use http::HeaderMap;
     use http::HeaderName;
@@ -2492,9 +2496,6 @@ mod tests {
     use crate::services::RouterResponse;
     use crate::services::connector::request_service::Request;
     use crate::services::connector::request_service::Response;
-    use crate::services::connector::request_service::TransportRequest;
-    use crate::services::connector::request_service::TransportResponse;
-    use crate::services::connector::request_service::transport;
     use crate::spec::operation_limits::OperationLimits;
 
     type JsonMap = serde_json_bytes::Map<ByteString, Value>;
@@ -3068,7 +3069,7 @@ mod tests {
                                         .unwrap();
                                     *http_request.headers_mut() = convert_http_headers(headers);
                                     let transport_request =
-                                        TransportRequest::Http(transport::http::HttpRequest {
+                                        TransportRequest::Http(HttpRequest {
                                             inner: http_request,
                                             debug: Default::default(),
                                         });
@@ -3185,7 +3186,7 @@ mod tests {
                                         context: Context::default(),
                                         connector: connector.into(),
                                         transport_result: Ok(TransportResponse::Http(
-                                            transport::http::HttpResponse {
+                                            HttpResponse {
                                                 inner: http_response.into_parts().0,
                                             },
                                         )),
