--- apollo-router/src/services/connector/request_service/transport.rs
@@ -1,2 +1,0 @@
-//! Transport implementations for Apollo Connectors
-pub(crate) mod http;
